refactor: re-organize file system-related modules
moving/improving existing functionality/documentation of several file
system-related code across modules
