+++ src/fs/mod.rs
@@ -1,0 +1,11 @@
+// SPDX-License-Identifier: MIT OR Apache-2.0
+// This file is part of Static Web Server.
+// See https://static-web-server.net/ for more information
+// Copyright (C) 2019-present Jose Quintana <joseluisq.net>
+
+//! A module that provides several facilities for working with files.
+//!
+
+pub(crate) mod meta;
+pub(crate) mod path;
+pub(crate) mod stream;
